Fixes deep toArray issue in Builder
If a `Collection` of `QueryableModel` instances is passed to `Builder::where`, if will deeply be converted to arrays, leading to the nested compileWhereQuery call to `throw new InvalidArrayValueException()` as the entry is now an array, and as it's associated, `count()` returns 0.
This fixes the issue by changing the `toArray` call to `all`.