Create array from labels when requesting validation.attributes
This hacks in some resemblance of form validation using netflex labels.

Netflex does not support or understand the naming syntax used by laravel. This makes it so that things like request validation fails to create acceptable translations due to the fact that the laravel request validator attempts to
fetch the key "validation.attributes" expecting an array.

Instead the LabelTranslator will create a new string key for the "validation.attributes" key, breaking translation of field names.

Since we can't add new keys in netflex and having to do so would be kind of painful regardless, we make a virtual array based on the "validation.values.fieldname." keys that are also created by the validator but would otherwise not be used.

NF-135 Make a small little netflex hack for validation.attributes