Only check that necessary services has been loaded
We don't need to have facades or to be fully booted for variables to be able to service real values.

This commit makes us check for the bare essentials before
attempting to contact netflex.

This lets us most likely use variables in boot methods.

NF-137 Variables in config

Fix typ

NF-137 Variables in config