Replace variable substitutes immediately before boot
We hook in and rebind config with a new repository that replaces the old one, but with all netflex variables
rebound to their correct values right before boot.

Instead of using the old method where we flattened the config array and set the new replacement values where applicable,
we now do a single recursive pass over the entire config array and create a new replacement array which we reinitialize
the config provider with.

We also skip using regex for this, as it is orders of magnitude slower than simple string searching.

NF-137 Variables in config