Fix Builder::random crashing if the query it randoms on produces no results
The random function performs an id search then shuffles and refetches random ids up to the selected amount.

However if the initial id query returns no results, the system tries to perform a new search for ids where the list is empty.

In this case the results should be an empty list. Therefore we just return an empty collection instead.