add handling of missing page controller
When building the page routes, each page's controller is invoked. In instances where a developer has registered a template with a controller in Netflex, and assigned it to a page, but not deployed the controller implementation, the page throws with a 500 everywhere.
This adds a fallback controller in such cases, which lets the route registration complete, and let's us show an error when the route is loaded instead.