Make Scheduler run the queue.worker's runNextJob instead of handle on job directly.
Laravel supports more queueable types than Jobs. The old implementation did not support them, as instead of running the worker, it just called the normal handle method on a job directly.

This had a bunch of negative side-effects, such as

* non-standard job handler names would not work
* scheduler would not emit events such as `JobProcessing`, `JobProcessed` or `JobFailed`
* queueable events would not work
* queueable event listeners would not work
* dependency injection in handle method would not work

This commit rewrites the scheduler handle endpoint in order to fake some aspects of the queue functionality that was unused in the previous implementation, as well as rewrites handle method itself so that it spawns a worker that runs a single job before quitting.

This way we're using laravel's built-in worker instead of our own implementation which makes scheduler based events work just like laravel events.

I don't think this approach will alter the behaviour of any currently written jobs that use the scheduler, as we were previously compliant with the most basic version of laravel jobs and therefore the new worker should operate just like our own, except better.